Move Message to slack package
